Modified the mapper name on the presentation layer and extracted the mapper from playone module.
