+++ mobile-ui/src/main/java/com/playone/mobile/ui/injection/module/LoginModule.kt
@@ -1,9 +1,9 @@
 package com.playone.mobile.ui.injection.module
 
 import com.playone.mobile.domain.interactor.auth.SignUpAndSignIn
+import com.playone.mobile.presentation.mapper.UserViewMapper
 import com.playone.mobile.presentation.onBoarding.LoginPlayoneContract
 import com.playone.mobile.presentation.onBoarding.LoginPlayonePresenter
-import com.playone.mobile.presentation.mapper.UserMapper
 import com.playone.mobile.ui.model.LoginViewModel
 import com.playone.mobile.ui.model.SignUpViewModel
 import dagger.Module
@@ -17,8 +17,8 @@ class LoginModule {
     }
 
     @Provides
-    internal fun provideLoginPresenter(signUpAndSignIn: SignUpAndSignIn, mapper: UserMapper)
-        : LoginPlayoneContract.Presenter = LoginPlayonePresenter(signUpAndSignIn, mapper)
+    internal fun provideLoginPresenter(signUpAndSignIn: SignUpAndSignIn, viewMapper: UserViewMapper)
+        : LoginPlayoneContract.Presenter = LoginPlayonePresenter(signUpAndSignIn, viewMapper)
 
     @Provides
     internal fun provideLoginViewModelFactory(presenter: LoginPlayoneContract.Presenter) =
