+++ mobile-ui/src/main/java/com/playone/mobile/ui/injection/module/mapper/DomainModule.kt
@@ -1,0 +1,6 @@
+package com.playone.mobile.ui.injection.module.mapper
+
+import dagger.Module
+
+@Module
+class DomainModule
