+++ mobile-ui/src/main/java/com/playone/mobile/ui/injection/module/mapper/PresentationModule.kt
@@ -1,0 +1,14 @@
+package com.playone.mobile.ui.injection.module.mapper
+
+import com.playone.mobile.presentation.mapper.UserViewMapper
+import dagger.Module
+import dagger.Provides
+import org.modelmapper.ModelMapper
+
+@Module
+class PresentationModule {
+
+    @Provides
+    internal fun proideUserMapper(modelMapper: ModelMapper) =
+        UserViewMapper(modelMapper)
+}
